pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "slm"
version = "0.1.0"
description = "From-scratch, production-style Small Language Model"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "torch>=2.0",
  "numpy>=1.26",
  "tqdm>=4.66",
  "pydantic>=2.6",
  "pyyaml>=6.0",
  "fastapi>=0.112",
  "uvicorn>=0.30",
  "rich>=13.7"
]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-q"
testpaths = ["tests"]
